Fallback support for globalize accessors
When using a locale specific accessor, we always want to get the raw
value of the localization, without using any fallback.

If we don't use this approach, and have an admin interface for
localizing a record, simply showing the edit screen and then saving
without changing anything will result in the fallback value being
inserted into the db. This makes it hard to find what translations are
outstanding, and unnecessarily polutes the database.